fixed missing id error

+++ app/src/main/java/com/example/birthdaywishes/pojo/Person.kt
@@ -8,7 +8,7 @@ import java.io.Serializable
 @Entity(tableName = "person_table")
 data class Person(val name: String, val birthday: Birthday, val phoneNumber: String) : Serializable{
 
-    @PrimaryKey(autoGenerate = true) var id: Int? = null
+    @PrimaryKey(autoGenerate = true) var id: Long? = null
 
     fun isDataValid() = name.isNotEmpty() &&
             phoneNumber.isNotEmpty() &&
